Adding Tags API, Refactoring API code layout
