+++ changedetectionio/tests/test_api_tags.py
@@ -1,67 +1,52 @@
 #!/usr/bin/env python3
 
-import time
 from flask import url_for
 from .util import live_server_setup, wait_for_all_checks
-
 import json
-import uuid
 
-
-def is_valid_uuid(val):
-    try:
-        uuid.UUID(str(val))
-        return True
-    except ValueError:
-        return False
-
-
-def test_setup(client, live_server, measure_memory_usage):
+def test_api_tags_listing(client, live_server, measure_memory_usage):
     live_server_setup(live_server)
-
-
-def test_api_tags(client, live_server, measure_memory_usage):
     api_key = live_server.app.config['DATASTORE'].data['settings']['application'].get('api_access_token')
+    tag_title = 'Test Tag'
 
-    # Make sure we start with no tags
+    # Get a listing
     res = client.get(
-        url_for("tags.tags_overview_page")
+        url_for("tags"),
+        headers={'x-api-key': api_key}
     )
-    assert b'No tags' in res.data
+    assert res.text.strip() == "{}", "Should be empty list"
+    assert res.status_code == 200
 
-    # Create a tag via API
     res = client.post(
-        "/api/v1/tag",
-        data=json.dumps({"title": "Test Tag"}),
-        headers={'content-type': 'application/json', 'x-api-key': api_key},
-        follow_redirects=True
+        url_for("tag"),
+        data=json.dumps({"title": tag_title}),
+        headers={'content-type': 'application/json', 'x-api-key': api_key}
     )
-
     assert res.status_code == 201
-    assert is_valid_uuid(res.json.get('uuid'))
-    tag_uuid = res.json.get('uuid')
+
+    new_tag_uuid = res.json.get('uuid')
 
     # List tags - should include our new tag
     res = client.get(
-        "/api/v1/tag",
+        url_for("tags"),
         headers={'x-api-key': api_key}
     )
     assert res.status_code == 200
-    assert tag_uuid in res.json
-    assert res.json[tag_uuid]['title'] == 'Test Tag'
-    assert res.json[tag_uuid]['notification_muted'] == False
+    assert new_tag_uuid in res.text
+    assert res.json[new_tag_uuid]['title'] == tag_title
+    assert res.json[new_tag_uuid]['notification_muted'] == False
 
     # Get single tag
     res = client.get(
-        f"/api/v1/tag/{tag_uuid}",
+        url_for("tag", uuid=new_tag_uuid),
         headers={'x-api-key': api_key}
     )
     assert res.status_code == 200
-    assert res.json['title'] == 'Test Tag'
+    assert res.json['title'] == tag_title
 
     # Update tag
     res = client.put(
-        f"/api/v1/tag/{tag_uuid}",
+        url_for("tag", uuid=new_tag_uuid),
         data=json.dumps({"title": "Updated Tag"}),
         headers={'content-type': 'application/json', 'x-api-key': api_key}
     )
@@ -70,7 +55,7 @@ def test_api_tags(client, live_server, measure_memory_usage):
 
     # Verify update worked
     res = client.get(
-        f"/api/v1/tag/{tag_uuid}",
+        url_for("tag", uuid=new_tag_uuid),
         headers={'x-api-key': api_key}
     )
     assert res.status_code == 200
@@ -78,7 +63,7 @@ def test_api_tags(client, live_server, measure_memory_usage):
 
     # Mute tag notifications
     res = client.get(
-        f"/api/v1/tag/{tag_uuid}?muted=muted",
+        url_for("tag", uuid=new_tag_uuid) + "?muted=muted",
         headers={'x-api-key': api_key}
     )
     assert res.status_code == 200
@@ -86,7 +71,7 @@ def test_api_tags(client, live_server, measure_memory_usage):
 
     # Verify muted status
     res = client.get(
-        f"/api/v1/tag/{tag_uuid}",
+        url_for("tag", uuid=new_tag_uuid),
         headers={'x-api-key': api_key}
     )
     assert res.status_code == 200
@@ -94,7 +79,7 @@ def test_api_tags(client, live_server, measure_memory_usage):
 
     # Unmute tag
     res = client.get(
-        f"/api/v1/tag/{tag_uuid}?muted=unmuted",
+        url_for("tag", uuid=new_tag_uuid) + "?muted=unmuted",
         headers={'x-api-key': api_key}
     )
     assert res.status_code == 200
@@ -102,16 +87,16 @@ def test_api_tags(client, live_server, measure_memory_usage):
 
     # Verify unmuted status
     res = client.get(
-        f"/api/v1/tag/{tag_uuid}",
+        url_for("tag", uuid=new_tag_uuid),
         headers={'x-api-key': api_key}
     )
     assert res.status_code == 200
     assert res.json['notification_muted'] == False
 
-    # Create a watch with the tag
+    # Create a watch with the tag and check it matches UUID
     test_url = url_for('test_endpoint', _external=True)
     res = client.post(
-        "/api/v1/watch",
+        url_for("createwatch"),
         data=json.dumps({"url": test_url, "tag": "Updated Tag", "title": "Watch with tag"}),
         headers={'content-type': 'application/json', 'x-api-key': api_key},
         follow_redirects=True
@@ -119,40 +104,40 @@ def test_api_tags(client, live_server, measure_memory_usage):
     assert res.status_code == 201
     watch_uuid = res.json.get('uuid')
 
-    # Verify tag is associated with watch
+    # Verify tag is associated with watch by name if need be
     res = client.get(
         url_for("watch", uuid=watch_uuid),
         headers={'x-api-key': api_key}
     )
     assert res.status_code == 200
-    assert tag_uuid in res.json.get('tags', [])
+    assert new_tag_uuid in res.json.get('tags', [])
 
     # Delete tag
     res = client.delete(
-        f"/api/v1/tag/{tag_uuid}",
+        url_for("tag", uuid=new_tag_uuid),
         headers={'x-api-key': api_key}
     )
     assert res.status_code == 204
 
     # Verify tag is gone
     res = client.get(
-        "/api/v1/tag",
+        url_for("tags"),
         headers={'x-api-key': api_key}
     )
     assert res.status_code == 200
-    assert tag_uuid not in res.json
+    assert new_tag_uuid not in res.text
 
     # Verify tag was removed from watch
     res = client.get(
-        f"/api/v1/watch/{watch_uuid}",
+        url_for("watch", uuid=watch_uuid),
         headers={'x-api-key': api_key}
     )
     assert res.status_code == 200
-    assert tag_uuid not in res.json.get('tags', [])
+    assert new_tag_uuid not in res.json.get('tags', [])
 
     # Delete the watch
     res = client.delete(
-        f"/api/v1/watch/{watch_uuid}",
+        url_for("watch", uuid=watch_uuid),
         headers={'x-api-key': api_key},
     )
     assert res.status_code == 204
